Convert uint to use int
Benefit of explicit does not show, I never get accidently scan wrong negative number into these struct.

However using uint requires several cast at places as interacting with golang gitlab sdk, most call use int as data type.

I hit the strange error when passing uint to GetGroup using the ID as int.

So just forget about it and use generic int instead.

+++ gitlab/projects/update-team-project.go
@@ -24,7 +24,7 @@ func UpdateTeamProject(git *gitlab.Client, p *gitlab.Project) {
 		u.CheckErr(err, "pSrv.ListProjectsGroups")
 		for _, pg := range projectGroups {
 			if strings.HasPrefix(pg.Name, "Team -") {
-				tp := TeamProjectNew(uint(pg.ID), uint(p.ID))
+				tp := TeamProjectNew(pg.ID, p.ID)
 				log.Printf("[DEBUG] TeamProjectNew %s\n", u.JsonDump(tp, "  "))
 			}
 		}
